refactor: reorganize OqtopusConfig import and implementation across modules
