+++ src/quri_parts_oqtopus/backend/estimation.py
@@ -9,7 +9,7 @@ from quri_parts.circuit import NonParametricQuantumCircuit
 from quri_parts.core.operator import Operator
 from quri_parts.openqasm.circuit import convert_to_qasm_str
 
-from quri_parts_oqtopus.backend import OqtopusConfig
+from quri_parts_oqtopus.backend.configuration import OqtopusConfig
 from quri_parts_oqtopus.rest import (
     ApiClient,
     Configuration,
