+++ src/quri_parts_oqtopus/backend/sampling.py
@@ -78,13 +78,11 @@ Examples:
 
 """
 
-import configparser
 import json
 import os
 import time
 from collections import Counter
 from datetime import datetime
-from pathlib import Path
 from typing import Any
 
 from quri_parts.backend import (
@@ -96,6 +94,7 @@ from quri_parts.backend import (
 from quri_parts.circuit import NonParametricQuantumCircuit
 from quri_parts.openqasm.circuit import convert_to_qasm_str
 
+from quri_parts_oqtopus.backend.configuration import OqtopusConfig
 from quri_parts_oqtopus.rest import (
     ApiClient,
     Configuration,
@@ -523,122 +522,6 @@ class OqtopusSamplingJob(SamplingJob):  # noqa: PLR0904
 
         """
         return self._job.to_str()
-
-
-class OqtopusConfig:
-    """A configuration information class for using OQTOPUS backend.
-
-    Args:
-        url: Base URL for OQTOPUS Cloud.
-        api_token: API token for OQTOPUS Cloud.
-
-    Raises:
-        ValueError: If ``url`` or ``api_token`` is None.
-
-    """
-
-    def __init__(self, url: str, api_token: str, proxy: str | None = None) -> None:
-        super().__init__()
-
-        if url is None:
-            msg = "url should not be None."
-            raise ValueError
-        self._url: str = url
-
-        if api_token is None:
-            msg = "api_token should not be None."
-            raise ValueError(msg)
-        self._api_token: str = api_token
-
-        self._proxy: str | None = proxy
-
-    @property
-    def url(self) -> str:
-        """Return the url.
-
-        Returns:
-            str: the url to access OCTOPUS Cloud.
-
-        """
-        return self._url
-
-    @property
-    def api_token(self) -> str:
-        """Return the API token.
-
-        Returns:
-            str: the API token to access OCTOPUS Cloud.
-
-        """
-        return self._api_token
-
-    @property
-    def proxy(self) -> str | None:
-        """Return the proxy.
-
-        Returns:
-            str | None: the proxy to access OCTOPUS Cloud.
-
-        """
-        return self._proxy
-
-    @staticmethod
-    def from_file(
-        section: str | None = "default", path: str | None = "~/.oqtopus"
-    ) -> "OqtopusConfig":
-        """Read configuration information from a file.
-
-        Args:
-            section: A :class:`OqtopusConfig` for circuit execution.
-            path: A path for config file.
-
-        Returns:
-            Configuration information :class:`OqtopusConfig` .
-
-        Examples:
-            The OQTOPUS configuration file describes configuration information for each
-            section. A section has a header in the form ``[section]``.
-            The default file path is ``~/.oqtopus`` and the default section name is
-            ``default``. Each section describes a setting in the format ``key=value``.
-            An example of a configuration file description is as below:
-
-            .. code-block::
-
-                [default]
-                url=<base URL>
-                api_token=<API token>
-
-                [sectionA]
-                url=<base URL>
-                api_token=<API token>
-
-                [sectioB]
-                url=<base URL>
-                api_token=<API token>
-                proxy=http://<proxy>:<port>
-
-            If ``sectionA`` settings are to be used, initialize
-            ``OqtopusSamplingBackend`` as follows
-
-            .. code-block::
-
-                backend = OqtopusSamplingBackend(OqtopusConfig.from_file("sectionA"))
-
-        """
-        if os.getenv("OQTOPUS_ENV") == "sse_container":
-            # This section is only for inside SSE container.
-            # Config is not needed in the container.
-            return OqtopusConfig(url="", api_token="")
-
-        path = Path(os.path.expandvars(path))
-        path = Path.expanduser(path)
-        parser = configparser.ConfigParser()
-        parser.read(path, encoding="utf-8")
-        return OqtopusConfig(
-            url=parser[section]["url"],
-            api_token=parser[section]["api_token"],
-            proxy=parser[section].get("proxy", None),
-        )
 
 
 class OqtopusSamplingBackend:
