+++ src/quri_parts_oqtopus/backend/sse.py
@@ -5,9 +5,10 @@ from pathlib import Path, PurePath
 
 from quri_parts.backend import BackendError
 
+from quri_parts_oqtopus.backend.configuration import OqtopusConfig
 from quri_parts_oqtopus.rest import ApiClient, Configuration, JobApi
 
-from .sampling import OqtopusConfig, OqtopusSamplingBackend, OqtopusSamplingJob
+from .sampling import OqtopusSamplingBackend, OqtopusSamplingJob
 
 
 class OqtopusSseBackend:
