+++ tests/quri_parts_oqtopus/backend/test_sampling.py
@@ -19,8 +19,8 @@ from pytest_mock.plugin import MockerFixture
 from quri_parts.backend import BackendError
 from quri_parts.circuit import QuantumCircuit
 
+from quri_parts_oqtopus.backend.configuration import OqtopusConfig
 from quri_parts_oqtopus.backend.sampling import (
-    OqtopusConfig,
     OqtopusSamplingBackend,
     OqtopusSamplingJob,
     OqtopusSamplingResult,
