+++ tests/quri_parts_oqtopus/backend/test_sse.py
@@ -25,9 +25,9 @@ from pytest_mock import MockerFixture
 from quri_parts.backend import BackendError
 
 from quri_parts_oqtopus.backend import (
-    OqtopusConfig,
     OqtopusSseBackend,
 )
+from quri_parts_oqtopus.backend.configuration import OqtopusConfig
 from quri_parts_oqtopus.rest import (
     JobsJobDef,
     JobsSubmitJobResponse,
